+++ MapFileDict/MapFileDict.csproj
@@ -58,6 +58,7 @@
     <Compile Include="MemMapStream.cs" />
     <Compile Include="MyClassThatRunsIn32and64Bit.cs" />
     <Compile Include="NativeMethods.cs" />
+    <Compile Include="OutOfProc.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
